requirements.txt
numpy
Pillow
matplotlib
seaborn
pandas
mpi4py
psutil
memory_profiler
pyyaml 